pyproject.toml
[tool.poetry]
name = "quenten"
version = "0.1.0"
description = ""
authors = ["Andy Rae <andy.rae@nottingham.ac.uk>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.5"
django-crispy-forms = "^1.14.0"
python-dotenv = "^0.20.0"

[tool.poetry.dev-dependencies]
pre-commit = "^2.19.0"
django-debug-toolbar = "^3.4.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
